followed srp for routes and cleand code principles
