+++ source_code/server/repository/user_repository.py
@@ -55,56 +55,8 @@ class UserRepository(IUser):
         return User(**data) if data else None
 
 
-    def remove_notification_keyword(self, userid, category_name: str, keyword=None):
-        if not keyword:
-            return self.__handle_category_removal(userid, category_name)
-        else:
-            return self.__handle_keyword_removal(userid, category_name, keyword)
 
-
-    def __handle_category_removal(self, userid: int, category_name: str):
-        preferences_list = self.__fetch_user_preferences(userid)
-        updated_preferences = [
-            category for category in preferences_list
-            if category['name'].lower() != category_name.lower()
-        ]
-
-        if len(updated_preferences) == len(preferences_list):
-            return {"success": False, "message": f"Category '{category_name}' not found"}, 400
-
-        self.__update_user_preferences(userid, updated_preferences)
-        return {"success": True, "message": f"Category '{category_name}' removed"}, 200
-
-
-    def __handle_keyword_removal(self, userid: int, category_name: str, user_input_keywords):
-        preferences_list = self.__fetch_user_preferences(userid)
-        updated_preferences = []
-        keyword_removed = False
-        
-        for category in preferences_list:
-            if category['name'].lower() == category_name.lower():
-                original_keywords_list = category.get('keywords', [])
-                updated_keywords = [
-                    keyword for keyword in original_keywords_list 
-                    if keyword.lower() not in  user_input_keywords
-                ]
-                if len(updated_keywords) != len(original_keywords_list):
-                    keyword_removed = True
-                category['keywords'] = updated_keywords
-            updated_preferences.append(category)
-
-        response = self.__handle_keyword_not_removed(user_input_keywords, category_name, keyword_removed)
-        if not response:
-            self.__update_user_preferences(userid, updated_preferences)
-            response = {
-                "success": True,
-                "message": f"Keyword '{user_input_keywords}' removed from category '{category_name}'"
-            }, 200
-        
-        return response 
-
-
-    def __update_user_preferences(self, userid, updated_preferences):
+    def update_user_preferences(self, userid, updated_preferences):
         update_query = update_user_preferences_by_id_query
         cursor_params = CursorDto(
             query=update_query,
@@ -115,7 +67,7 @@ class UserRepository(IUser):
         db.execute_query(cursor_params)
 
 
-    def __fetch_user_preferences(self, userid):
+    def fetch_user_preferences(self, userid):
         query = get_user_preferences_by_id_query
         cursor_params = CursorDto(query=query, params=(userid,), fetch_one=True)
         result: dict = db.execute_query(cursor_params)
@@ -129,11 +81,3 @@ class UserRepository(IUser):
             return json.loads(preferences)
         except json.JSONDecodeError:
             return []
-
-
-    def __handle_keyword_not_removed(self, keyword, category_name, keyword_removed):
-        if not keyword_removed:
-            return {
-                "success": False,
-                "message": f"Keyword '{keyword}' not found in category '{category_name}'"
-            }, 400
