+++ source_code/server/utils/serializers.py
@@ -1,0 +1,2 @@
+def serialize_articles(article_list):
+    return [article.__dict__ for article in article_list]
